right aligned floating point output

+++ chapter1/main.c
@@ -5,7 +5,7 @@ fahrenheit to celsius converter
 */
 int main(int argc, char **argv)
 {
-	int fahr, celsius;
+	float fahr, celsius;
 	int lower, upper, step;
 	lower = 0;
 	upper = 300;
@@ -13,7 +13,7 @@ int main(int argc, char **argv)
 	fahr = lower;
 	while(fahr <= upper){
 		celsius = 5 * (fahr - 32) / 9;
-		printf("%3d\t%6d\n", fahr, celsius);
+		printf("%3.0f\t%6.1f\n", fahr, celsius);
 		fahr += step;
 	}
 	return 0;
